Move EXPORT_INTEREST_STATUS to common/constants
